Add files via upload
This is a simple project which creates a GUI inside which you can set timer in seconds. And when the condition satisfies it plays sound. In this project playsound,threading, and tkinter library are used, to make code successfully work. You can also add some more functionality like stop timer,resume timer etc. But I have made it basic.